globals
This commit adds the 'WebSocket', 'Blob' and 'crypto' globals for Deno
and the browser. It adds 'requestAnimationFrame' and
'cancelAnimationFrame', as well as 'ResizeObserver' for the browser.

It removes 'name' and 'event' for the browser. This is because these are
more commonly used as variable and parameter names, and having them
implicitly declared can prevent JSLint catching some bugs.