import.meta
This commit adds support for the import.meta object. It is used to
resolve the address of files relative to the current module. This makes
it much easier to use Workers, which are otherwise resolved relative to
the page rather than the module:

    new Worker("/path/to/my_worker.js");                    // awkward
    new Worker(new URL("./my_worker.js", import.meta.url)); // clunky
    new Worker(import.meta.resolve("./my_worker.js"));      // better

Workers are available in the browser and Deno. They allow JavaScript
programs to distribute their work over multiple CPU cores, which is
becoming increasingly important.

This commit also repairs JSLint's handling of the dynamic import
function, allowing it to be used in expression position.