a?.b.c
Previously, JSLint was permitting the optional chaining operator only at the
conclusion of a compound refinement:

    a.b?.c           // OK
    a?.b.c           // Not OK

However, the optional chaining operator short circuits at the first nullish
value, meaning that 'a?.b.c' is safe even when 'a' is nullish.

However, JSLint still prohibits expressions with adjacent ?. because that is
difficult to reason about:

    a?.b?.c          // Not OK

In adding this feature, I noticed that JSLint was permitting the use of ?.
in some invocations, for example:

    "x"?.repeat(4)   // OK

This is now fixed.