import.meta.main
The 'import.meta.main' boolean lets a module adjust its behavior depending on
whether it is run as an entrypoint or imported as a module. For example, a
module might demonstrate its own correctness when run as a program, but have no
side effects during importation. This is a good thing.