_
This commit renames the reserved word 'ignore' to '_', as suggested by Dale
Schumacher. This is a very breaking change, but a necessary one.

'_' has some advantages over 'ignore'. For one, it is shorter. For
another, it is obviously special, whereas the word 'ignore' could be a
legitimate identifier in many programs. Lastly, it is an accepted convention in
many languages to use '_' to indicate an identifier that should not be bound.

Why the word 'ignore' was originally used is unclear, but it may be have been
to avoid a conflict with the Lodash library that was popular at the time.
Lodash used the global variable '_' to hold its functions.