globalThis
Historically, the global object has been known by many names: window, self,
global, this. However, the name of the global object is now specified by the
JavaScript standard and so there is no longer any reason for this proliferation
of aliases.

This commit removes 'window' and 'self' from the allowed browser globals. The
only way to access the global object by name is via 'globalThis', and such
mischief is only permitted under the new "global" option.