Reverse stacks
Due to the lack of array methods that run in reverse, it is
unnecessarily awkward to iterate over the function and block stacks when
they are manipulated with push/pop. With very minimal performance
impact, push/pop has been replaced with shift/unshift, making it
possible to use the short circuiting 'every' method to search up the
stack.