Block scope
Previously, JSLint placed all variable names into function scopes,
meaning it was impossible to reuse a variable name in sibling scopes:

    if (a) {
        const b = 42;
        ...
    } else {
        const b = 1729;  // WARNING!
        ...
    }

The common workaround, used within JSLint itself, was to declare a
mutable variable and assign to it in whichever block happened to
execute:

    let b;
    if (a) {
        b = 42;
        ...
    } else {
        b = 1729;
        ...
    }

The problem with this workaround is that the variable is accessible to
more scopes than is necessary, and we lose the clarity gained by
using const over let.

This commit makes blocks, rather than functions, the primary scopes.

The original behavior was necessary to accomodate var, an obsolete
alternative to const and let that is no longer supported by JSLint.